feat: Plugin architecture
